move getHours
